seeding is complete

+++ back-end/db/seeds/seed.ts
@@ -1,15 +1,17 @@
 import {db} from '../connection'
-import { InsertManyOptions, MongooseError, InferSchemaType} from 'mongoose'
+import mongoose, { InsertManyOptions, MongooseError, InferSchemaType} from 'mongoose'
 import { User} from '../schemas/userschema'
 import { Item } from '../schemas/itemschema'
 
 export const seed = async (users: any, items: any) => {
-    await db
+
+    if (mongoose.connection.collections["User"]) {
     await User.collection.drop()
+    }
+
     await Item.collection.drop()
     const insertedUsers = await User.insertMany(users)
 
     const insertedItems = await Item.insertMany(items)
-    console.log(insertedUsers)
-    console.log(insertedItems)
+    console.log('hello')
 }
